+++ RecordParser.Test/FileReaderTest.cs
@@ -1,11 +1,10 @@
 ﻿using FluentAssertions;
 using RecordParser.Builders.Reader;
-using RecordParser.Extensions;
+using RecordParser.Extensions.FileReader;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
-using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Xunit;
