+++ RecordParser/Extensions/FileReader/QuotedRow.cs
@@ -3,7 +3,7 @@ using System.Buffers;
 using System.Collections.Generic;
 using System.IO;
 
-namespace RecordParser.Extensions
+namespace RecordParser.Extensions.FileReader
 {
     public static partial class Exasd
     {
@@ -62,7 +62,7 @@ namespace RecordParser.Extensions
                 return totalRead;
             }
 
-            public IEnumerable<Memory<char>> TryReadLine()
+            public IEnumerable<ReadOnlyMemory<char>> TryReadLine()
             {
                 int Peek() => i < bufferLength ? buffer[i] : -1;
